fix: add activity tab (wip)
