+++ app/components/icons/arrow.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Arrow = () => {
+  return (
+    <svg width="14" height="12" viewBox="0 0 14 12" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <path d="M8.5 1.5L13 6L8.5 10.5M12.25 6L1 6" stroke="white" stroke-opacity="0.4" stroke-width="1.6875" stroke-linecap="round" stroke-linejoin="round"/>
+    </svg>
+  );
+};
+
+export default Arrow;
